refactor: 패키지 구조 mvc 패턴 구조로 변경
