+++ src/main/java/me/namuhuchutong/lotto/LottoApplication.java
@@ -1,5 +1,10 @@
 package me.namuhuchutong.lotto;
 
+import me.namuhuchutong.lotto.controller.LottoController;
+import me.namuhuchutong.lotto.domain.generator.RandomNumberGenerator;
+import me.namuhuchutong.lotto.ui.InputView;
+import me.namuhuchutong.lotto.ui.OutputView;
+
 public class LottoApplication {
 
     public static void main(String[] args){
