+++ src/main/java/me/namuhuchutong/lotto/domain/LottoCount.java
@@ -1,4 +1,4 @@
-package me.namuhuchutong.lotto;
+package me.namuhuchutong.lotto.domain;
 
 
 public class LottoCount {
