+++ src/main/java/me/namuhuchutong/lotto/domain/Number.java
@@ -1,4 +1,4 @@
-package me.namuhuchutong.lotto;
+package me.namuhuchutong.lotto.domain;
 
 
 import java.util.Objects;
